Revert "Remove clone dedupe"
This reverts commit 6b0742ce0aa5860dda9cb8a40a7e47f466cdceb6.
